+++ docs/src/utils/CruiseLogo.js
